Move Shape and DataType into src/keras_format
