+++ src/engine/container.ts
@@ -14,8 +14,9 @@ import {Scalar, serialization, Tensor, tidy, util} from '@tensorflow/tfjs-core';
 
 import {getUid} from '../backend/state';
 import {NotImplementedError, RuntimeError, ValueError} from '../errors';
+import {PyJsonDict, Shape} from '../keras_format/types';
 import {deserialize as deserializeLayer} from '../layers/serialization';
-import {Kwargs, NamedTensorMap, PyJsonDict, Shape} from '../types';
+import {Kwargs, NamedTensorMap} from '../types';
 import * as generic_utils from '../utils/generic_utils';
 import {convertTsToPythonic} from '../utils/serialization_utils';
 import * as types_utils from '../utils/types_utils';
